requirements.txt
asgiref==3.11.1
Django==6.0.3
gunicorn==25.3.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pygame-ce==2.5.6
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
